pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "strubloid-translate"
version = "1.2.0"
description = "Real-time audio transcription and translation using OpenAI Whisper"
authors = [{ name = "Strubloid" }]
requires-python = ">=3.8"
dependencies = [
    "torch",
    "numpy",
    "ffmpeg-python",
    "tqdm",
    "more-itertools",
    "openai>=1.0.0",
    "numba",
    "typing-extensions",
    "python-dotenv>=0.21.0",
    "openai-whisper @ git+https://github.com/openai/whisper.git"
]
